In Regards to last boss update push
Gitkraken pulled a fast one and made of an auto stash of my changes. The last push includes completed movement for boss, boss is now killable, player will re spawn at the beginning if killed by boss or enemy and discovered problem with Phealth which is too ingrained in our code to get rid of. Phealth Problem: Phealth for some reason will not update between files instead it will default back to its original value inside combat.h file which means we can't have the player take damage from both enemy and boss since they will have different values. instead what we have now is the player can deal damage to enemies in enemies.cpp and to the boss in bossenemy.cpp and the enemies deal damage to the player in enemies,cpp and the boss in bossenemy.cpp. This means if we want to track the players health we need to do it in both enemies.cpp and or bossenemy.cpp.